requirements.txt
aiogram>=3.13
sqlalchemy>=2.0
aiosqlite>=0.20
aiohttp>=3.10
matplotlib>=3.10
apscheduler>=3.10
python-dotenv>=1.0
